# public/swagger.yaml
openapi: 3.0.0
info:
  title: Genouest Account Manager (My) API
  description: # todo
  version: 1.4.32
...
paths:
...
  /ssh/{id}:
    tags:
      - ssh
    parameters:
      - $ref: '#/components/parameters/uid'
    get:
      summary: Generate SSH key
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-UserNotFound'
        '500':
          $ref: '#/components/responses/500-ServerError'
  /ssh/{id}/putty:
    tags:
      - ssh
    parameters:
      - $ref: '#/components/parameters/uid'
    get:
      summary: Download Putty SSH key
      responses:
        '200':
          $ref: '#/components/responses/200-SSHDownload'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-NotFound'
  /ssh/{id}/private:
    tags:
      - ssh
    parameters:
      - $ref: '#/components/parameters/uid'
    get:
      summary: Download private SSH key
      responses:
        '200':
          $ref: '#/components/responses/200-SSHDownload'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-NotFound'
  /ssh/{id}/public:
    tags:
      - ssh
    parameters:
      - $ref: '#/components/parameters/uid'
    get:
      summary: Download public SSH key
      responses:
        '200':
          $ref: '#/components/responses/200-SSHDownload'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-NotFound'
  /tags:
    tags:
      - tags
    get:
      summary: Get all tags
      responses:
        '200':
          description: OK - Send tags
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      type: string
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
  /tags/{kind}/{id}:
    tags:
      - tags
    parameters:
      - in: paths
        name: kind
        required: true
        schema:
          type: string
        description: Either 'group' or 'user'
      - $ref: '#/components/parameters/uid'
    post:
      summary: Get all tags associated with a user or a group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          $ref: '#/components/responses/200-SendConfirmation'
        '401':
          $ref: '#/components/responses/401-NotAuthorized'
        '403':
          $ref: '#/components/responses/403-UnsanitaryId'
        '404':
          $ref: '#/components/responses/404-NotFound'
...
components:
  parameters:
    uid:
      in: paths
      name: id
      required: true
      schema:
        type: string
      description: User uid (username)
...
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
      example:
        message: 'User session not found'
...
  responses:
    200-SendConfirmation:
      description: OK
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
...
    200-SSHDownload:
      description: SSH key downloaded
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary
...
    401-NotAuthorized:
      description: Authorization information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
...
    403-UnsanitaryId:
      description: Input id is not sanitized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
...
    404-NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    404-UserNotFound:
      description: A user with the specified uid was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
    500-ServerError:
      description: An unexpected server error has occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'